requirements.txt
flask_jwt_extended
flask_sqlalchemy
flask_marshmallow
flask_cors
flask_migrate
flask_script
flask_bcrypt
flask_mail
flask_uploads
flask_restful
Flask-login
Werkzeug
requests
pyJWT
Flask-mail 
Flask-login
